Clean up create files utility
